fb_app/requirements.txt
pywebview==4.4.1
selenium==4.16.0
beautifulsoup4==4.12.2
webdriver-manager==4.0.1 